Preparations for integration of TPTriagMeshBuilder
